fix: Introduce an "application admin" user to run migrations
This mean we have three database users:

1. The PostgreSQL "superadmin" ("postgres"): Only used for initial
   database creation.
2. The "application admin" ("dbadmin"): Only used for applying
   schema migrations.
3. The "application user" ("dbuser"): Used for accessing the database.

Each user require a separate password in a "secrets" file.

We also ensure that a *single* image is built for both the "backend" and
"backend-init" services by tagging it as "bird-ringing-backend:local"
(for now).

The backend entrypoint script is also updated, as is the documentation.