fix: Ensure that persistent volumes are up-to-date on startup
The persistent volumes were not buing updated with the most recent data
from the build process, if they already existed. This caused issues
when there were changes in the data that needed to be reflected in the
volumes.

The solutien is to build the static files into the containers during the
build process (we did this before too, but we no longer designate the
target directory as a "volume"), and then copy them into the persistent
volumes during the container startup (this is a new step). This way,
the volumes will always have the latest data when the containers are
started, even if the volumes already exist.

This neccessitates mounting the volumes as read/write in the "backend"
and "frontend" containers (previously, we mounted them as read-only).